Update page.js
Hi I was adding some custom delete functionality and noticed that when you delete a single record on the last page it doesn't redirect to the start page. Added a simple change to make sure we are redirected to a start page when there are no items to display on current page.